rename the incorrectly named package "nl.tudelft.jewel" to "nl.tudelft.bejeweled.jewel" <bug
