+++ src/main/java/nl/tudelft/bejeweled/jewel/ExplosivePowerUp.java
@@ -1,4 +1,4 @@
-package nl.tudelft.jewel;
+package nl.tudelft.bejeweled.jewel;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -6,16 +6,15 @@ import java.util.List;
 import javafx.scene.Group;
 import javafx.scene.Node;
 import nl.tudelft.bejeweled.sprite.ExplosiveSprite;
-import nl.tudelft.bejeweled.sprite.HyperSprite;
 import nl.tudelft.bejeweled.sprite.JewelSprite;
 import nl.tudelft.bejeweled.sprite.Sprite;
 
-public class HyperPowerUp extends JewelPowerUp {
+public class ExplosivePowerUp extends JewelPowerUp {
 
 	private JewelSprite powerUpSprite;
-	public HyperPowerUp(Jewel baseJewel) {
+	public ExplosivePowerUp(Jewel baseJewel) {
 		super(baseJewel);
-		this.powerUpSprite = new HyperSprite(super.getSprite().getxPos(), super.getSprite().getyPos()); 
+		this.powerUpSprite = new ExplosiveSprite(super.getSprite().getxPos(), super.getSprite().getyPos()); 
 	}
 	
 	@Override
@@ -51,10 +50,5 @@ public class HyperPowerUp extends JewelPowerUp {
 		super.implode(sceneGroup);
 		powerUpSprite.implode(sceneGroup);
 		
-	}
-	
-	@Override
-	public boolean isHyper(){
-		return true;
 	}
 }
