+++ src/test/java/nl/tudelft/bejeweled/board/BoardTest.java
@@ -10,9 +10,9 @@ import org.junit.Before;
 import org.junit.Test;
 
 import javafx.collections.ObservableList;
+import nl.tudelft.bejeweled.jewel.BasicJewel;
+import nl.tudelft.bejeweled.jewel.Jewel;
 import nl.tudelft.bejeweled.sprite.SpriteStore;
-import nl.tudelft.jewel.BasicJewel;
-import nl.tudelft.jewel.Jewel;
 
 /**
  * Various tests for the Board class.
