+++ src/test/java/nl/tudelft/bejeweled/sprite/JewelTest.java
@@ -3,8 +3,8 @@ package nl.tudelft.bejeweled.sprite;
 import org.junit.Before;
 import org.junit.Test;
 
-import nl.tudelft.jewel.BasicJewel;
-import nl.tudelft.jewel.Jewel;
+import nl.tudelft.bejeweled.jewel.BasicJewel;
+import nl.tudelft.bejeweled.jewel.Jewel;
 
 import static org.junit.Assert.assertEquals;
